moved 'Streaming' folder to 'Expenditures', renamed fundingSource to fundingSources
