+++ src/modules/constants.ts
@@ -159,3 +159,5 @@ export const ALLDOMAINS_DOMAIN_SELECTION = {
 };
 
 export const SMALL_TOKEN_AMOUNT_FORMAT = '0.00000...';
+
+export const LOCAL_STORAGE_EXPENDITURE_TYPE_KEY = 'expediture-type';
