+++ src/modules/dashboard/components/ExpenditurePage/Streaming/Streaming.css
@@ -1,4 +1,5 @@
 @value paddingRightLeft: 54px;
+@value paddingBottom: 54px;
 
 .circlePlusIcon {
   display: flex;
@@ -14,7 +15,7 @@
 }
 
 .container {
-  padding: 0 paddingRightLeft;
+  padding: 0 paddingRightLeft paddingBottom;
 }
 
 .header {
