+++ src/modules/pages/components/ExpenditurePage/constants.ts
@@ -4,10 +4,10 @@ import { nanoid } from 'nanoid';
 import { toFinite } from 'lodash';
 import { defineMessages } from 'react-intl';
 
-import { newFundingSource } from '~dashboard/ExpenditurePage/ExpenditureSettings/Streaming/constants';
 import { newRecipient } from '~dashboard/ExpenditurePage/Payments/constants';
 import { initalRecipient } from '~dashboard/ExpenditurePage/Split/constants';
 import { initalMilestone } from '~dashboard/ExpenditurePage/Staged/constants';
+import { newFundingSource } from '~dashboard/ExpenditurePage/Streaming/constants';
 
 import { ExpenditureTypes } from './types';
 
@@ -56,7 +56,7 @@ export const initialValues = {
     ],
   },
   streaming: {
-    fundingSource: [newFundingSource],
+    fundingSources: [newFundingSource],
   },
 };
 
@@ -137,5 +137,3 @@ export const validationSchema = yup.object().shape({
     }),
   }),
 });
-
-export const EXPENDITURE_TYPE_KEY = 'expediture-type';
